notification/requirements.txt
uvicorn
fastapi
python-telegram-bot
python-multipart
ddtrace